DirectXフレームプロファイラとCS+MS描画パイプライン導入
DirectXのCPU/GPUフレーム計測用FrameProfilerを追加し、Time EditorでFPS/処理時間を可視化。FixFPSをリファクタし、FPS/デルタタイムの取得・設定を強化。チャンク描画をCompute Shader＋Mesh Shader方式に刷新し、Chunkクラスのリソース管理を最適化。Test2Phaseで武器モデルのボーン追従描画とImGui調整機能を追加。その他、MapManagerの半径初期値やTestAnimation/Playerの細部修正も含む。

+++ NobleEngine/src/application/GameObjects/Character/Player/Player.cpp
@@ -42,7 +42,7 @@ void Player::Update(int32_t cameraID)
 	{
 	}
 
-	previousHP_ = HP_;
+	previousHP_ = static_cast<float>(HP_);
 
 	// 移動系入力処理
 	UpdateInput(cameraID);
